hv: sched: add vcpu_affinity[] in vm_config to support vcpu assignment
To support CPU sharing, we need known how to assign vcpus to each pcpu.
Add this vcpu_affinity[] for each VM to indicate the assignment policy.
With it, we also may not set pcpu_bitmap as we can get it from
vcpu_affinity[] and cpu_num.

This patch also add some sanitize check of vcpu_affinity[]. Here are
some rules:
  1) only one bit can be set for each vcpu_affinity of vcpu.
  2) two vcpus in same VM cannot be set with same vcpu_affinity.
  3) vcpu_affinity cannot be set to the pcpu which used by pre-launched VM.

Signed-off-by: Jason Chen CJ <jason.cj.chen@intel.com>
Signed-off-by: Yu Wang <yu1.wang@intel.com>
Signed-off-by: Shuo A Liu <shuo.a.liu@intel.com>